api_gateway/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
httpx>=0.24.0
pyjwt>=2.6.0
strawberry-graphql>=0.171.1
python-multipart>=0.0.6
